TST/CLN: Break out groupby function tests (#34073)
